+++ src/Tree/Node.php
@@ -35,7 +35,7 @@ abstract class Node
     private $attributes = [];
 
     /**
-     * @var Node[]
+     * @var array[]
      */
     private $children = [];
 
@@ -101,6 +101,14 @@ abstract class Node
     }
 
     /**
+     * Check if this is a null object implementation
+     */
+    public function isNull(): bool
+    {
+        return false;
+    }
+
+    /**
      * Set a custom attribute on node
      *
      * @param string $name  Name of attribute
@@ -141,21 +149,23 @@ abstract class Node
     /**
      * Set a child node
      */
-    public function addChild(string $name, Node $child): void
+    public function addChild(string $name, Node $node): void
     {
-        $this->children[$name] = $child;
+        $this->children[] = [$name, $node];
     }
 
     /**
      * Get child node
      */
-    public function getChild(string $name): ?Node
+    public function getChild(string $name): Node
     {
-        if (!$this->hasChild($name)) {
-            return null;
+        foreach ($this->children as list($nodeName, $node)) {
+            if (strcasecmp($nodeName, $name) == 0) {
+                return $node;
+            }
         }
 
-        return $this->children[$name];
+        return new NullNode;
     }
 
     /**
@@ -163,16 +173,26 @@ abstract class Node
      */
     public function hasChild(string $name): bool
     {
-        return isset($this->children[$name]);
+        foreach ($this->children as list($nodeName, $node)) {
+            if (strcasecmp($nodeName, $name) == 0) {
+                return true;
+            }
+        }
+
+        return false;
     }
 
     /**
      * Get registered child nodes
      *
-     * @return Node[]
+     * @return iterable & Node[]
      */
-    public function getChildren(): array
+    public function getChildren(string $name = ''): iterable
     {
-        return $this->children;
+        foreach ($this->children as list($nodeName, $node)) {
+            if (!$name || strcasecmp($nodeName, $name) == 0) {
+                yield $node;
+            }
+        }
     }
 }